chore(directories): moved factory files to temporary directory to avoid merge conflict
