+++ core/src/main/kotlin/com/tdt4240/group3/model/temporaryFactory/CapitalFactory.kt
@@ -1,14 +1,15 @@
-package com.tdt4240.group3.model.ecs.entities
+package com.tdt4240.group3.model.temporaryFactory
 
 import com.badlogic.ashley.core.Engine
+import com.tdt4240.group3.model.components.CapitalComponent
 import com.tdt4240.group3.model.ecs.components.CityComponent
 import com.tdt4240.group3.model.ecs.components.PositionComponent
 import com.tdt4240.group3.model.ecs.components.TeamComponent
 import ktx.ashley.entity
 import ktx.ashley.with
 
-class CityFactory(private val engine: Engine): Factory<CityConfig> {
-    override fun createEntity(config: CityConfig) = engine.entity {
+class CapitalFactory(private val engine: Engine): Factory<CapitalConfig> {
+    override fun createEntity(config: CapitalConfig) = engine.entity {
         with<CityComponent> {
             this.name = config.name
             this.baseProduction = config.baseProduction
@@ -21,5 +22,6 @@ class CityFactory(private val engine: Engine): Factory<CityConfig> {
         with<TeamComponent> {
             this.team = config.team
         }
+        with<CapitalComponent> { } //TODO("Add an origin in capital component")
     }
 }
