+++ core/src/main/kotlin/com/tdt4240/group3/model/temporaryFactory/CityFactory.kt
@@ -1,15 +1,14 @@
-package com.tdt4240.group3.model.ecs.entities
+package com.tdt4240.group3.model.temporaryFactory
 
 import com.badlogic.ashley.core.Engine
-import com.tdt4240.group3.model.components.CapitalComponent
 import com.tdt4240.group3.model.ecs.components.CityComponent
 import com.tdt4240.group3.model.ecs.components.PositionComponent
 import com.tdt4240.group3.model.ecs.components.TeamComponent
 import ktx.ashley.entity
 import ktx.ashley.with
 
-class CapitalFactory(private val engine: Engine): Factory<CapitalConfig> {
-    override fun createEntity(config: CapitalConfig) = engine.entity {
+class CityFactory(private val engine: Engine): Factory<CityConfig> {
+    override fun createEntity(config: CityConfig) = engine.entity {
         with<CityComponent> {
             this.name = config.name
             this.baseProduction = config.baseProduction
@@ -22,6 +21,5 @@ class CapitalFactory(private val engine: Engine): Factory<CapitalConfig> {
         with<TeamComponent> {
             this.team = config.team
         }
-        with<CapitalComponent> { } //TODO("Add an origin in capital component")
     }
 }
