+++ core/src/main/kotlin/com/tdt4240/group3/model/temporaryFactory/Factory.kt
@@ -1,4 +1,4 @@
-package com.tdt4240.group3.model.ecs.entities
+package com.tdt4240.group3.model.temporaryFactory
 
 import com.badlogic.ashley.core.Entity
